+++ SistemaIntegrado/SistemaIntegrado/Funcionalidad/Crear/Factory/IPerfilFactory.cs
@@ -5,6 +5,6 @@ namespace SistemaIntegradoAlertas.Funcionalidad.Crear.Factory
 {
     public interface IPerfilFactory
     {
-        Perfil CrearPerfil(string nombre, string correo, string tipoCedula, uint cedula, uint celular, string contrasena, params object[] extras);
+        Perfil CrearPerfil(string nombre, string correo, string tipoCedula, uint cedula, long celular, string contrasena, params object[] extras);
     }
 }
