+++ SistemaIntegrado/SistemaIntegrado/Funcionalidad/Crear/Factory/OperadorFactory.cs
@@ -7,13 +7,13 @@ namespace SistemaIntegrado.Funcionalidad.Crear.Factory
 {
     public class OperadorFactory : IPerfilFactory
     {
-        public Perfil CrearPerfil(string nombre, string correo, string tipoCedula, uint cedula, uint celular, string contrasena, params object[] extras)
+        public Perfil CrearPerfil(string nombre, string correo, string tipoCedula, uint cedula, long celular, string contrasena, params object[] extras)
         {
             // Para Operador: extras[0] = int numOperador, extras[1] = List<Alerta> alertas
             int numOperador = extras.Length > 0 && extras[0] is int ? (int)extras[0] : 1;
             List<Alerta> alertas = extras.Length > 1 && extras[1] is List<Alerta> ? (List<Alerta>)extras[1] : new List<Alerta>();
             
-            return new Operador(nombre, correo, (int)celular, tipoCedula, (int)cedula, contrasena, numOperador, alertas);
+            return new Operador(nombre, correo, celular, tipoCedula, (int)cedula, contrasena, numOperador, alertas);
         }
     }
 }
