+++ SistemaIntegrado/SistemaIntegrado/Funcionalidad/Crear/Factory/PacienteFactory.cs
@@ -6,7 +6,7 @@ namespace SistemaIntegrado.Funcionalidad.Crear.Factory
 {
     public class PacienteFactory : IPerfilFactory
     {
-        public Perfil CrearPerfil(string nombre, string correo, string tipoCedula, uint cedula, uint celular, string contrasena, params object[] extras)
+        public Perfil CrearPerfil(string nombre, string correo, string tipoCedula, uint cedula, long celular, string contrasena, params object[] extras)
         {
             // Para Paciente: extras[0] = Historia_Clinica, extras[1] = float latitud, extras[2] = float longitud
             Historia_Clinica historia = extras.Length > 0 && extras[0] is Historia_Clinica ? (Historia_Clinica)extras[0] : 
@@ -14,7 +14,7 @@ namespace SistemaIntegrado.Funcionalidad.Crear.Factory
             float latitud = extras.Length > 1 && extras[1] is float ? (float)extras[1] : 0.0f;
             float longitud = extras.Length > 2 && extras[2] is float ? (float)extras[2] : 0.0f;
             
-            return new Ciudadano(nombre, correo, (int)celular, tipoCedula, (int)cedula, contrasena, historia, latitud, longitud);
+            return new Ciudadano(nombre, correo, celular, tipoCedula, (int)cedula, contrasena, historia, latitud, longitud);
         }
     }
 }
